try to forward that gpg agent

+++ bashrc.d/500-gpg.sh
@@ -7,8 +7,16 @@ if [ -e "$HOME/.gpg-agent-info" ] ; then
     rm -f $HOME/.gitconfig
     ln -s $HOME/.homesick/repos/dotfiles/home/.gitconfig $HOME/.gitconfig
   fi
-elif [ -L $HOME/.gitconfig ] || [ $HOME/.homesick/repos/dotfiles/home/.gitconfig -nt $HOME/.gitconfig ] ; then
-  rm -f $HOME/.gitconfig
-  cp $HOME/.homesick/repos/dotfiles/home/.gitconfig $HOME/.gitconfig
-  git config --global commit.gpgsign false
+elif [ -e $HOME/.gnupg/S.remote-gpg-agent ] ; then
+  export GPG_AGENT_INFO=$HOME/.gnupg/S.remote-gpg-agent.$$
+  rm -f $GPG_AGENT_INFO
+  mv $HOME/.gnupg/S.remote-gpg-agent $GPG_AGENT_INFO
+fi
+
+if [ -z "$GPG_AGENT_INFO" ] ; then
+  if [ -L $HOME/.gitconfig ] || [ $HOME/.homesick/repos/dotfiles/home/.gitconfig -nt $HOME/.gitconfig ] ; then
+    rm -f $HOME/.gitconfig
+    cp $HOME/.homesick/repos/dotfiles/home/.gitconfig $HOME/.gitconfig
+    git config --global commit.gpgsign false
+  fi
 fi
